+++ crates/symworx-loadsym/src/nutrition/calories.rs
@@ -1,4 +1,4 @@
-// Copyright (c) 2026 PalEm Dynamics LLC
+// Copyright (c) 2026 Nathaniel T. Berry
 // Licensed under the Apache License, Version 2.0.
 
 //! Calorie and energy expenditure calculations: BMR (Mifflin-St Jeor + obesity